back-end/requirements.txt
dj-database-url
Django
django-database-url
gunicorn
psycopg2
sqlparse
djangorestframework
django-cors-headers
requests
django-filter
